separate router, add babel alias
