+++ .babelrc
@@ -30,7 +30,16 @@
     ],
     "@babel/plugin-proposal-nullish-coalescing-operator",
     "@babel/plugin-proposal-do-expressions",
-    "@babel/plugin-proposal-function-bind"
+    "@babel/plugin-proposal-function-bind",
+    [
+      "module-resolver",
+      {
+        "root": "./application",
+        "alias": {
+          "#application": "./application",
+        }
+      }
+    ]
   ],
   "env": {
     "development": {
