--- application/components/Loading.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <div>
-    <div>
-      <div />
-    </div>
-  </div>
-);
