--- application/components/home2/index.js
@@ -1,6 +1,0 @@
-import universal from 'react-universal-component';
-import Loading from '../loading/loading';
-
-export default universal(() => import('./home'), {
-  loading: Loading
-});
