+++ application/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Router from './router';
+
+const App = () => (
+  <div>
+    Header
+    <Router />
+    Footer
+  </div>
+);
+
+export default App;
