+++ application/views/home/index.js
@@ -1,5 +1,5 @@
 import universal from 'react-universal-component';
-import Loading from '../loading/loading';
+import Loading from '#application/components/loading/loading';
 
 export default universal(() => import('./home'), {
   loading: Loading
