+++ client/entry.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { hydrate } from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import App from '../application/components/App2';
+import App from '../application';
 import store from '../application/store/store';
 
 const render = (App) => hydrate(
@@ -20,8 +20,8 @@ const theApp = () => (
 );
 
 if (process.env.NODE_ENV === 'development' && module.hot) {
-  module.hot.accept('../application/components/App2.js', () => {
-    const theApp = require('../application/components/App2').default; // eslint-ignore-line
+  module.hot.accept('../application/', () => {
+    const theApp = require('../application/').default; // eslint-ignore-line
     render(theApp);
   });
 }
