+++ package.json
@@ -64,6 +64,7 @@
     "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
+    "babel-plugin-module-resolver": "^3.2.0",
     "babel-plugin-universal-import": "^4.0.0",
     "babel-watch": "^7.0.0",
     "commitizen": "^3.1.1",
@@ -72,6 +73,7 @@
     "cz-conventional-changelog": "^2.1.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb": "^17.1.0",
+    "eslint-import-resolver-babel-module": "^5.1.0",
     "eslint-plugin-flowtype": "^3.9.1",
     "eslint-plugin-import": "^2.17.3",
     "eslint-plugin-jsx-a11y": "^6.2.1",
