+++ server/entry.js
@@ -4,7 +4,7 @@ import { flushChunkNames } from 'react-universal-component/server';
 import flushChunks from 'webpack-flush-chunks';
 import { Provider } from 'react-redux';
 import { StaticRouter } from 'react-router-dom';
-import App from '../application/components/App2';
+import App from '../application';
 import store from '../application/store/store';
 
 export default ({ clientStats }) => (req, res) => {
